+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [8.3.2] - 2020-07-26
+### Alpha feature sendInvoice
+ - Added sendInvoice, answerShippingQuery, answerPreCheckoutQuery for testing - [#119](https://github.com/windkh/node-red-contrib-telegrambot/issues/119)
+ 
 ## [8.3.1] - 2020-07-26
 ### Fixed
  - Fixed has response behavior - [#115](https://github.com/windkh/node-red-contrib-telegrambot/issues/115)